Move code writing Lua data to tables into their own files
