+++ src/CMakeLists.txt
@@ -51,6 +51,7 @@ if (WITH_LUA)
         flex-table-column.cpp
         flex-lua-geom.cpp
         flex-lua-index.cpp
+        flex-write.cpp
         geom-transform.cpp
         lua-utils.cpp
         output-flex.cpp
